Coordenadas de Centro: paso tercer test

+++ db/migrate/021_add_latitud_longitud_centros.rb
@@ -1,7 +1,7 @@
 Sequel.migration do
   up do
-    add_column :centros, :latitude, BigDecimal
-    add_column :centros, :longitude, BigDecimal
+    add_column :centros, :latitude, Float
+    add_column :centros, :longitude, Float
   end
 
   down do
